Fix cover image
Apparently, the images were tagged as executable and I'm not sure if that had to be the problem why it wasn't working. So I changed the permissions to be the same as the cover given in the template cover and now it magically works! All the other images I intended on using also seemed to be executable so I changed it to read-only. However the images between chapters seemed to display fine even after I changed it to read-only. Nevertheless, I accomplished my goal of displaying the cover image. Never thought that knowing how to change file permissions would help me out in a situation like this! I used this site http://www.maclife.com/article/columns/terminal_101_file_permissions to quickly learn how to do it.