Add chapter images
Adds chapter image before each chapter. Resized images to a suitable size for ereaders. For some reason, when I added the images to the manifest in content.opf, they stopped displaying and would break the page so for the images and it's corresponding captions to display properly, they must not be added to the manifest.